refactor: component 구조 개선
