+++ js/App.js
@@ -1,3 +1,2 @@
 import './Router.js';
 
-
