+++ js/components/container/AppLayout.js
@@ -1,0 +1,12 @@
+import Header from '../template/Header.js';
+import CreateElement from '../../lib/CreateElement.js';
+
+const AppLayout = ({ children }) => (
+  CreateElement('div',
+    {},
+    Header(),
+    children(),
+  )
+);
+
+export default AppLayout;
