+++ js/components/template/kanban/TodoTitle.js
@@ -1,4 +1,4 @@
-import CreateElement from '../../lib/CreateElement.js';
+import CreateElement from '../../../lib/CreateElement.js';
 
 const TodoTitle = ({ name }) => {
   const dom = CreateElement('h2');
