+++ js/constant/STRINGS.js
@@ -1,4 +1,5 @@
 export default {
   titlePostfix: '\'s Todo List',
   nameValidator: '이름은 최소 2글자 이상이어야 합니다.',
+  todoInputPlaceHoler: '할 일을 입력해주세요.',
 };
